Implement reward drop weights
Weights are arbitrary positive numbers which are dynamically normalized
to a probability [0, 1] for each set of weights. (The sum of each set is
1.0)

Rolls are implemented as a biased distribution of probabilities.
See https://stackoverflow.com/a/479299 and
http://en.wikipedia.org/wiki/Uniform_distribution

Unfortunately there is quite a bit of looping but the size of each array
is only O(50) which really only applies to cards. So, worst case, in the
unlikely case that all N rewards are cards, the framerate may hiccup for
a moment. I think this is ok because 1) the situation is extremely
unlikely and 2) I don't think it's actually enough loops to cause missed
update/render ticks (except on very, very low end computers, maybe).